// apps/web/src/lib/api/generated/models/assistantRunDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DeepRef API
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantRunDtoError } from './assistantRunDtoError.ts';

export interface AssistantRunDto {
	/** @nullable */
	answer_message_id?: string | null;
	/** @nullable */
	completed_at?: string | null;
	conversation_id: string;
	created_at: string;
	error: AssistantRunDtoError;
	id: string;
	/** @nullable */
	plan_id?: string | null;
	project_id: string;
	status: string;
	trigger_message_id: string;
	updated_at: string;
}